+++ lib/overcommit/errors.rb
@@ -1,0 +1,3 @@
+class NotAGitRepoError < StandardError
+  # Nothing to see here
+end